+++ src/store/actions/shipmentActions.js
@@ -16,6 +16,14 @@ export const DELETING_PACKAGE = 'DELETING_PACKAGE';
 export const DELETING_PACKAGE_SUCCESSFUL = 'DELETING_PACKAGE_SUCCESSFUL';
 export const DELETING_PACKAGE_FAILURE = 'DELETING_PACKAGE_FAILURE';
 
+export const ADDING_PACKAGE = 'ADDING_PACKAGE';
+export const ADDING_PACKAGE_SUCCESSFUL = 'ADDING_PACKAGE_SUCCESSFUL';
+export const ADDING_PACKAGE_FAILURE = 'ADDING_PACKAGE_FAILURE';
+
+export const SELECTED_PRODUCT = 'SELECTED_PRODUCT';
+
+export const DELETE_SELECTED = 'DELETE_SELECTED';
+
 export const addShipment = (trackingNumber, productId) => dispatch => {
 	const trackingRequest = {
 		trackingNumber,
@@ -32,6 +40,25 @@ export const addShipment = (trackingNumber, productId) => dispatch => {
 			dispatch({ type: ADDING_SHIPMENT_FAILURE, payload: err.data }),
 		);
 };
+
+export const addPackage = packageArr => dispatch => {
+	const products = { products: packageArr };
+	console.log(products);
+	dispatch({ type: ADDING_PACKAGE });
+	axios
+		.post('/packaging/preview', products)
+		.then(res =>
+			axios
+				.post('/packaging/addweb', res.data)
+				.then(res =>
+					dispatch({ type: ADDING_PACKAGE_SUCCESSFUL, payload: res.data }),
+				),
+		)
+		.catch(err =>
+			dispatch({ type: ADDING_PACKAGE_FAILURE, payload: err.data }),
+		);
+};
+
 
 export const getShipments = () => dispatch => {
 	dispatch({ type: GETTING_SHIPMENTS });
@@ -67,4 +94,12 @@ export const deletePackage = uuid => dispatch => {
 		.catch(err =>
 			dispatch({ type: DELETING_PACKAGE_FAILURE, payload: err.data }),
 		);
+};
+
+export const selectProduct = product => dispatch => {
+	dispatch({ type: SELECTED_PRODUCT, payload: product });
+};
+
+export const deleteSelectedProduct = uuid => dispatch => {
+	dispatch({ type: DELETE_SELECTED, payload: uuid });
 };
